requirements.txt
appdirs==1.4.4
Django>=3.0.6,<3.1.0
djangorestframework>=3.11.0,<4.1.0
